+++ yaralyzer/decoding/bytes_decoder.py
@@ -15,8 +15,8 @@ from rich.text import Text
 from yaralyzer.bytes_match import BytesMatch
 from yaralyzer.config import YaralyzerConfig
 from yaralyzer.decoding.decoding_attempt import DecodingAttempt
-from yaralyzer.encoding_detection.encoding_assessment import EncodingAssessment
 from yaralyzer.encoding_detection.character_encodings import ENCODING, ENCODINGS_TO_ATTEMPT
+from yaralyzer.encoding_detection.encoding_assessment import EncodingAssessment
 from yaralyzer.encoding_detection.encoding_detector import EncodingDetector
 from yaralyzer.helpers.dict_helper import get_dict_key_by_value
 from yaralyzer.helpers.rich_text_helper import CENTER, DECODING_ERRORS_MSG, NO_DECODING_ERRORS_MSG
@@ -25,7 +25,7 @@ from yaralyzer.output.decoding_attempts_table import (DecodingTableRow, assessme
 from yaralyzer.output.rich_console import console
 from yaralyzer.util.logging import log
 
-# Messages used in the table to show true vs. false (a two element array can be indexed by booleans)
+# A 2-tuple that can be indexed by booleans of messages used in the table to show true vs. false
 WAS_DECODABLE_YES_NO = [NO_DECODING_ERRORS_MSG, DECODING_ERRORS_MSG]
 
 # Multiply chardet scores by 100 (again) to make sorting the table easy
@@ -34,7 +34,6 @@ SCORE_SCALER = 100.0
 
 class BytesDecoder:
     def __init__(self, bytes_match: BytesMatch, label: Optional[str] = None) -> None:
-        """Instantiated with _bytes as the whole stream; :bytes_seq tells it how to pull the bytes it will decode"""
         self.bytes_match = bytes_match
         self.bytes = bytes_match.surrounding_bytes
         self.label = label or bytes_match.label
@@ -51,6 +50,7 @@ class BytesDecoder:
         self.encoding_detector = EncodingDetector(self.bytes)
 
     def print_decode_attempts(self) -> None:
+        """Print the DecodingAttemptsTable."""
         console.line(2)
         self._print_decode_attempt_subheading()
 
@@ -108,7 +108,7 @@ class BytesDecoder:
         console.print(panel, justify=CENTER)
 
     def _track_decode_stats(self) -> None:
-        "Track stats about successful vs. forced vs. failed decode attempts"
+        """Track stats about successful vs. forced vs. failed decode attempts"""
         for decoding in self.decodings:
             if decoding.failed_to_decode:
                 self.was_match_undecodable[decoding.encoding] += 1
@@ -120,11 +120,15 @@ class BytesDecoder:
                 self.was_match_force_decoded[decoding.encoding] += 1
 
     def _row_from_decoding_attempt(self, decoding: DecodingAttempt) -> DecodingTableRow:
+        """
+        Create a DecodingAttemptTable row from a DecodingAttempt.
+        If the decoding result is a duplicate of a previous decoding, replace the decoded text
+        with "same output as X" where X is the previous encoding that gave the same result.
+        """
         assessment = self.encoding_detector.get_encoding_assessment(decoding.encoding)
         plain_decoded_string = decoding.decoded_string.plain
         sort_score = assessment.confidence * SCORE_SCALER
 
-        # Replace the decoded text with a "same output as X" where X is the encoding that gave the same result
         if plain_decoded_string in self.decoded_strings.values():
             encoding_with_same_output = get_dict_key_by_value(self.decoded_strings, plain_decoded_string)
             display_text = Text('same output as ', style='color(66) dim italic')